requirements.txt
Flask==2.3.2
flask-sqlalchemy~=3.0.5
flask-migrate~=4.0.4
Werkzeug~=2.3.6
python-dotenv~=1.0.0